+++ include/main_runtime_state.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "../src/settings.h"
+
 struct MainRuntimeState {
     bool bootReady = false;
     unsigned long bootReadyDeadlineMs = 0;
@@ -10,5 +12,7 @@ struct MainRuntimeState {
     unsigned long v1ConnectedAtMs = 0;
     bool wifiAutoStartDone = false;
     unsigned long lastLoopUs = 0;
+    bool v1InfoSavedThisConnect = false;  // true once v1_info.json written after current connect
+    DrivingMode lastDrivingMode = DrivingMode::Normal;  // detect mode changes in main loop
 };